+++ a_uction/src/main/java/com/example/a_uction/config/ElasticSearchConfig.java
@@ -1,6 +1,7 @@
 package com.example.a_uction.config;
 
 import org.elasticsearch.client.RestHighLevelClient;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.elasticsearch.client.ClientConfiguration;
 import org.springframework.data.elasticsearch.client.RestClients;
@@ -12,10 +13,13 @@ import org.springframework.data.elasticsearch.repository.config.EnableElasticsea
 @EnableElasticsearchRepositories
 public class ElasticSearchConfig extends AbstractElasticsearchConfiguration {
 
+    @Value("${es.host}")
+    private String host;
+
     @Override
     public RestHighLevelClient elasticsearchClient() {
         ClientConfiguration clientConfiguration = ClientConfiguration.builder()
-                .connectedTo("3.35.38.11:9200")
+                .connectedTo(host)
                 .build();
         return RestClients.create(clientConfiguration).rest();
     }
