+++ a_uction/src/main/java/com/example/a_uction/model/auctionSearch/dto/ItemNameSearchCondition.java
@@ -1,0 +1,15 @@
+package com.example.a_uction.model.auctionSearch.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.domain.Sort;
+
+@Getter
+@Setter
+public class ItemNameSearchCondition {
+
+    private String itemName;
+    private String sortProperties;
+    private Sort.Direction direction;
+
+}
