+++ a_uction/src/test/java/com/example/a_uction/controller/auction/AuctionControllerTest.java
@@ -23,7 +23,7 @@ import com.example.a_uction.model.auction.dto.AuctionDto;
 import com.example.a_uction.model.auction.entity.AuctionEntity;
 import com.example.a_uction.model.user.entity.UserEntity;
 import com.example.a_uction.security.jwt.JwtProvider;
-import com.example.a_uction.service.auction.AuctionSearchService;
+import com.example.a_uction.service.search.AuctionSearchService;
 import com.example.a_uction.service.auction.AuctionService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.FileInputStream;
