+++ elastic-search.http
@@ -22,3 +22,5 @@ GET http://localhost:8081/search/auctions/description/matches?description=애플
 ###description contains
 GET http://localhost:8081/search/auctions/description/contains?description=iphone&page=0&size=20
 
+###item name
+GET localhost:8081/search/auctions/itemName/startWith?itemName=아이폰&direction=DESC&page=0&size=20
